LayoutTests/platform/wpe/imported/w3c/web-platform-tests/html/semantics/document-metadata/the-meta-element/color-scheme/meta-color-scheme-insert-expected.txt


FAIL Initial color-scheme assert_equals: expected (string) "normal" but got (undefined) undefined
FAIL Inserted meta color-scheme applies assert_equals: expected (string) "light" but got (undefined) undefined
FAIL Inserted meta color-scheme before existing in head applies assert_equals: expected (string) "dark" but got (undefined) undefined
